global関数のクラス内関数化、クラスのカプセル化、MApChipData.h の追加(#3)
