+++ Thirteen/Thirteen.vcxproj
@@ -106,7 +106,7 @@
     </ClCompile>
     <Link>
       <SubSystem>NotSet</SubSystem>
-      <AdditionalDependencies>Winmm.lib;D3d9.lib;D3dX9.lib;dinput8.lib;DXGUID.LIB;%(AdditionalDependencies)</AdditionalDependencies>
+      <AdditionalDependencies>Winmm.lib;D3d9.lib;D3dX9.lib;dinput8.lib;DXGUID.LIB;d3dx9.h;D3dxtex.h;%(AdditionalDependencies)</AdditionalDependencies>
     </Link>
   </ItemDefinitionGroup>
   <ItemDefinitionGroup Condition="'$(Configuration)|$(Platform)'=='Release|Win32'">
@@ -171,6 +171,7 @@
     <ClInclude Include="GameClear.h" />
     <ClInclude Include="GameOver.h" />
     <ClInclude Include="Graphics.h" />
+    <ClInclude Include="MapchipData.h" />
     <ClInclude Include="Input.h" />
     <ClInclude Include="MainGame.h" />
     <ClInclude Include="Meteorite.h" />
