+++ Thirteen/Thirteen.vcxproj.filters
@@ -132,6 +132,9 @@
     <ClInclude Include="SceneSwitch.h">
       <Filter>ソース ファイル\Scene</Filter>
     </ClInclude>
+    <ClInclude Include="MapchipData.h">
+      <Filter>ソース ファイル\Texture</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <None Include="cpp.hint" />
